pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinelscan"
version = "1.0.0"
description = "SentinelScan: Static + Online file scanner (hash, YARA, libmagic, PE, heuristics, VT/OPSWAT)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Sergio Castilla Roncel", email = "sergiocastillaroncel@gmail.com" }]
dependencies = [
  "rich>=13.7.0",
  "yara-python>=4.5.0",
  "pefile>=2023.2.7",
  "python-magic>=0.4.27",
  "requests>=2.32.0"
]

[project.urls]
Homepage = "https://github.com/sergiodev2/SentinelScan"
Repository = "https://github.com/sergiodev2/SentinelScan"

[project.scripts]
sentinelscan = "sentinelscan.cli:main"
